+++ pages/site_sidebar.php
@@ -40,13 +40,26 @@
     <?php } ?>
 
     <?php if ($rq_sec['admin']=='1' OR $rq_sec['compliance']=='1'){ ?>
-      <li><a href="activos.php"><i class="fa fa-archive"></i> <span>Activos</span></a></li>
-      <li><a href="controles.php"><i class="fa fa-retweet"></i> <span>Controles</span></a></li>
-      <li><a href="iso27k.php"><i class="fa fa-crosshairs"></i> <span>ISO 27001</span></a></li>
-      <li><a href="iso9k.php"><i class="fa fa-crosshairs"></i> <span>ISO 9001</span></a></li>
-      <li><a href="bcra.php"><i class="fa fa-crosshairs"></i> <span>BCRA</span></a></li>
-      <li><a href="mejoras.php"><i class="fa fa-refresh"></i> <span>Mejora Continua</span></a></li>
-      <li><a href="riesgos.php"><i class="fa fa-flash"></i> <span>Riesgos</span></a></li>
+        <li><a href="activos.php"><i class="fa fa-archive"></i> <span>Activos</span></a></li>
+        <li><a href="controles.php"><i class="fa fa-retweet"></i> <span>Controles</span></a></li>
+        <li><a href="iso27k.php"><i class="fa fa-crosshairs"></i> <span>ISO 27001</span></a></li>
+        <li><a href="iso9k.php"><i class="fa fa-crosshairs"></i> <span>ISO 9001</span></a></li>
+        <li><a href="bcra.php"><i class="fa fa-crosshairs"></i> <span>BCRA</span></a></li>
+        <li><a href="mejoras.php"><i class="fa fa-refresh"></i> <span>Mejora Continua</span></a></li>
+        <li class="treeview">
+            <a href="#">
+                <i class="fa fa-eye"></i> <span>Auditorías</span>
+                <span class="pull-right-container">
+                <i class="fa fa-angle-left pull-right"></i>
+                </span>
+            </a>
+            <ul class="treeview-menu">
+                <li><a href="aud_ente.php"><i class="fa fa-building"></i> Entes Auditores</a></li>
+                <li><a href="aud_instancia.php"><i class="fa fa-clone"></i> Instancias</a></li>
+            </ul>
+        </li>    
+
+        <li><a href="riesgos.php"><i class="fa fa-flash"></i> <span>Riesgos</span></a></li>
     <?php }?>
     <?php if ($rq_sec['admin']=='1' OR $rq_sec['proy']=='1' OR $rq_sec['admin_proy']=='1'){
           echo '<li><a href="proyectos.php"><i class="fa fa-list"></i> <span>Proyectos</span></a></li>';